Component Page Consolidating
