+++ src/components/FlavorToDrink/CocktailList/CocktailList.jsx
@@ -1,6 +1,6 @@
 import "./CocktailList.module.css";
-import * as cocktailsAPI from "../../utilities/cocktails-api";
-import CocktailListItem from "../CocktailListItem/CocktailListItem";
+import * as cocktailsAPI from "../../../utilities/cocktails-api";
+import CocktailListItem from "../../FlavorToDrink/CocktailListItem/CocktailListItem";
 import { useEffect, useState } from "react";
 
 export default function CocktailList(props) {
