+++ src/components/MoodToMix/MixtapeList/MixtapeList.jsx
@@ -1,0 +1,6 @@
+export default function MixtapeList(props) {
+
+    return (
+      props.mixtapes.map(mixtape => <p>{mixtape.name}</p>)
+    )
+}
