+++ src/pages/FlavourPage/FlavourPage.jsx
@@ -1,6 +1,6 @@
 import style from "./FlavourPage.module.css"
 import Header from "../../components/Header/Header"
-import FlavourList from "../../components/FlavourList/FlavourList"
+import FlavourList from "../../components/FlavorToDrink/FlavourList/FlavourList"
 
 export default function FlavourPage(props) {
     return (
